Cast some types before printing
Explicitly cast mode_t and pid_t to integer types before printing them.
Fixes build problems in OpenSolaris 2008.11.

Monotone-Parent: 852e12918d82970cf8c9dcca3a54211c2f9f2955
Monotone-Revision: 5e0143d5f920b25d552efdd093ddebc2b16bd730

Monotone-Author: jmmv@NetBSD.org
Monotone-Date: 2009-01-18T15:49:02
Monotone-Branch: org.NetBSD.atf.src

+++ tests/atf/atf-c/t_fs.c
@@ -1,7 +1,7 @@
 /*
  * Automated Testing Framework (atf)
  *
- * Copyright (c) 2007, 2008 The NetBSD Foundation, Inc.
+ * Copyright (c) 2007, 2008, 2009 The NetBSD Foundation, Inc.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -666,7 +666,7 @@ ATF_TC_BODY(eaccess, tc)
         const int experr = atf_user_is_root() ? t->rerror : t->uerror;
 
         printf("\n");
-        printf("File mode     : %04o\n", t->fmode);
+        printf("File mode     : %04o\n", (unsigned int)t->fmode);
         printf("Access mode   : 0x%02x\n", t->amode);
 
         ATF_REQUIRE(chmod(atf_fs_path_cstring(&p), t->fmode) != -1);
